Add valueOf Method in DataType Class
